Added two new arguments
Added sub_clust_cut_off which is the quartile of the data at which a cluster's average correlation, if below said threshold, will then be broken up.

agglo_cut_off is the quartile cut off of the data, at which point it determines what correlation is required for the algorithm to stop including genes in clusters.

To make this code work cur_corr_storage is a new varaiable that holds 1000 correlation numbers. If for 1000 correlation changes, the minimum correlation has not changed, we use this variable to break from the while loop.